Remove experimental features and consolidate data management
