--- chrome-extension/manifest.json
@@ -1,23 +1,0 @@
-{
-  "manifest_version": 3,
-  "name": "ADHD Tools Sync",
-  "description": "Sync ADHD Tools data across your Chrome devices using chrome.storage.sync.",
-  "version": "1.0",
-  "icons": {
-    "16": "icon16.svg",
-    "48": "icon48.svg",
-    "128": "icon128.svg"
-  },
-  "action": {
-    "default_popup": "popup.html",
-    "default_title": "ADHD Tools Sync"
-  },
-  "permissions": ["storage", "activeTab", "scripting"],
-  "content_scripts": [
-    {
-      "matches": ["https://jobellet.github.io/ADHDtools/*"],
-      "js": ["content.js"],
-      "run_at": "document_idle"
-    }
-  ]
-}
