+++ index.html
@@ -10,6 +10,7 @@
     <script src="app.js" defer></script>
     <script src="data-manager.js" defer></script>
     <script src="gemini.js" defer></script>
+    <script src="gemini-settings.js" defer></script>
     <script src="pomodoro.js" defer></script>
     <script src="eisenhower.js" defer></script>
     <script src="day-planner.js" defer></script>
@@ -19,7 +20,6 @@
     <script src="focus-mode.js" defer></script>
     <script src="reward-system.js" defer></script>
     <script src="calendar-tool.js" defer></script>
-    <script src="data-management.js" defer></script>
     <script src="routine.js" defer></script>
     <script src="script-loader.js" defer></script>
     <script src="i18n.js" defer></script>
@@ -29,7 +29,6 @@
         <div class="container">
             <h1 id="site-title" data-i18n="site-title"><i class="fas fa-brain"></i> ADHD Tools Hub</h1>
             <p data-i18n="tagline">Interactive tools to help manage ADHD symptoms and improve productivity</p>
-            <button id="open-colab-backend">Open Colab Backend</button>
             <span id="current-time-display" class="current-time"></span>
         </div>
     </header>
