--- package.json
@@ -1,12 +1,0 @@
-{
-  "name": "adhdtools-extension-helper",
-  "version": "1.0.0",
-  "description": "Utility scripts for ADHD Tools",
-  "scripts": {
-    "share-extension": "node share-extension.js"
-  },
-  "dependencies": {
-    "archiver": "^6.0.0",
-    "qrcode-terminal": "^0.12.0"
-  }
-}
