+++ script-loader.js
@@ -1,10 +1,14 @@
 // Update index.html to include the new JavaScript files
 document.addEventListener('DOMContentLoaded', function() {
-    // Add script tags for calendar integration and data management
-    const calendarScript = document.createElement('script');
-    calendarScript.src = 'calendar-integration.js';
-    document.body.appendChild(calendarScript);
-    
+    // Load Google Calendar integration only if credentials are provided
+    const clientId = localStorage.getItem('gcalClientId');
+    const apiKey = localStorage.getItem('gcalApiKey');
+    if (clientId && apiKey) {
+        const calendarScript = document.createElement('script');
+        calendarScript.src = 'calendar-integration.js';
+        document.body.appendChild(calendarScript);
+    }
+
     // Avoid loading data-manager.js twice. If it's already
     // present in the document (e.g., included in the page head), don't
     // inject it again as re-running the script would recreate the
@@ -14,11 +18,11 @@ document.addEventListener('DOMContentLoaded', function() {
         dataManagerScript.src = 'data-manager.js';
         document.body.appendChild(dataManagerScript);
     }
-    
+
     // Add celebration container for rewards
     const celebrationContainer = document.createElement('div');
     celebrationContainer.id = 'celebration-container';
     document.body.appendChild(celebrationContainer);
-    
+
     console.log('Additional scripts and elements loaded');
 });
